+++ src/main/java/com/github/jkschoen/jsma/model/AlbumStats.java
@@ -12,7 +12,7 @@ import javax.xml.bind.annotation.XmlRootElement;
 public class AlbumStats extends Stats {
 	
 	@XmlElement(name="id")
-	private Integer id;
+	private Long id;
 	
 	@XmlElement(name="Key")
 	private String key;
@@ -23,7 +23,7 @@ public class AlbumStats extends Stats {
 	public AlbumStats() {
 	}
 	
-	public AlbumStats(Integer id, String key, Integer bytes, 
+	public AlbumStats(Long id, String key, Integer bytes, 
 			Integer hits, Integer large, Integer medium, 
 			Integer original, Integer small, Integer x2Large, 
 			Integer x3Large, Integer xLarge) {
@@ -35,14 +35,14 @@ public class AlbumStats extends Stats {
 	/**
 	 * The id for this image.
 	 */
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
 	/**
 	 * Sets the id for this image.
 	 */
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
